Added in tracking of an optimization variable
Allows you to add an optimization variable as an equation (z = c’ * x)
where z is your linear objective function,  c’ is an arbitrary
coefficient vector, and x represents the variables in your network. The
only change was to add a variable in the input query to keep track of
which one z is. I still need to add a way for this information to get
passed on to the tableau or cost function manager.